Bottom-up version of rod cutting. Refer book P366.
There is a big flaw in this version.
For function bottom_up_rod_cut(x, y, z), the first parameter
"rodValinLength" has never been called inside. Instead, the pointer 'p' to the
global array of rod revenue in diff length is used. That's why
 array revinLength size calculation is still correct.
But this flaw does not take side effects over the result.